% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLMARKERS.R
\docType{data}
\name{ALLMARKERS}
\alias{ALLMARKERS}
\title{Markers for Codestral}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{marker}{The marker.}
\item{decription}{The description of the marker.}
}
}
\usage{
ALLMARKERS
}
\description{
Markers for Codestral
}
\keyword{internal}
