% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allow_detect_package.R
\name{allow_detect_package}
\alias{allow_detect_package}
\title{Allow codestral to detect package environment}
\usage{
allow_detect_package(x = TRUE)
}
\arguments{
\item{x}{A boolean indicating that the user allows the detection of a
package environment. Defaulted to \code{TRUE}.}
}
\value{
\code{0} invisible.
}
\description{
Allow codestral to detect package environment
}
\details{
If set to \code{TRUE}, when codestral is used in a folder with a \code{DESCRIPTION}
file, all R files in the current folder and its subfolders are included
in the prompt.
}
