% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestral.R
\name{codestral}
\alias{codestral}
\title{Fill in the middle with Codestral}
\usage{
codestral(
  prompt,
  codestral_api_key = Sys.getenv(x = "R_CODESTRAL_APIKEY"),
  mistral_api_key = Sys.getenv(x = "R_MISTRAL_APIKEY"),
  fim_model = Sys.getenv(x = "R_CODESTRAL_FIM_MODEL"),
  chat_model = Sys.getenv(x = "R_CODESTRAL_CHAT_MODEL"),
  mamba_model = Sys.getenv(x = "R_MAMBA_CHAT_MODEL"),
  temperature = as.integer(Sys.getenv(x = "R_CODESTRAL_TEMPERATURE")),
  max_tokens = Sys.getenv(x = "R_CODESTRAL_MAX_TOKENS"),
  suffix = ""
)
}
\arguments{
\item{prompt}{The prompt to complete.}

\item{codestral_api_key}{The API key to use for accessing Codestral. Defaults to the value of the
\code{R_CODESTRAL_APIKEY} environment variable.}

\item{mistral_api_key}{The API key to use for accessing Mamba. Defaults to the value of the
\code{R_MISTRAL_APIKEY} environment variable.}

\item{fim_model}{The model to use for fill-in-the-middle. Defaults to the value of the
\code{R_CODESTRAL_FIM_MODEL} environment variable.}

\item{chat_model}{The model to use for chat with Codestral. Defaults to the value of the
\code{R_CODESTRAL_CHAT_MODEL} environment variable.}

\item{mamba_model}{The model to use for chat with Mamba. Defaults to the value of the
\code{R_MAMBA_CHAT_MODEL} environment variable.}

\item{temperature}{The temperature to use. Defaults to the value of the \code{R_CODESTRAL_TEMPERATURE}
environment variable.}

\item{max_tokens}{The maximum number of tokens to generate. Defaults to the value of the
\code{R_CODESTRAL_MAX_TOKENS} environment variable.}

\item{suffix}{The suffix to use. Defaults to an empty string.}
}
\value{
A character string containing the completed text.
}
\description{
This function completes a given prompt using the Codestral API. It supports different models
for fill-in-the-middle, chat with Codestral, and chat with Mamba. The function relies on
environment variables for some parameters.
}
\examples{
\dontrun{
codestral(prompt = "Once upon a time")
}
}
