% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestral_init.R
\name{codestral_init}
\alias{codestral_init}
\title{Initialize codestral}
\usage{
codestral_init(
  apikey = NULL,
  fim_model = "codestral-latest",
  chat_model = "codestral-latest",
  mamba_model = "open-codestral-mamba",
  temperature = 0,
  max_tokens = 100
)
}
\arguments{
\item{apikey}{Your codestral API key.}

\item{fim_model}{The model to use for FIM.}

\item{chat_model}{The model to use for chat.}

\item{mamba_model}{The model to use for mamba.}

\item{temperature}{The temperature to use.}

\item{max_tokens}{The maximum number of tokens to generate.}
}
\description{
Initialize codestral
}
