% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_dialog.R
\name{compile_dialog}
\alias{compile_dialog}
\title{Analyses a prompt to re-buid the dialog}
\usage{
compile_dialog(prompt)
}
\arguments{
\item{prompt}{The prompt to analyse. A vector of strings.}
}
\value{
A list with the chatter (Codestral or Codestral Mamba) and the dialog in a data.frame whith columns \code{role} and \code{content}.
}
\description{
Analyses a prompt to re-buid the dialog
}
