% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_current_script.R
\name{complete_current_script}
\alias{complete_current_script}
\title{Fill in the middle or complete}
\usage{
complete_current_script()
}
\value{
A character vector containing the two parts of the script.
}
\description{
This function splits the current script into two parts: the part before the cursor and the part after the cursor.
}
