% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_mode.R
\name{debug_mode}
\alias{debug_mode}
\title{Set debug mode}
\usage{
debug_mode(debug = TRUE)
}
\arguments{
\item{debug}{A logical value. If \code{TRUE}, the debug mode is activated.}
}
\value{
Invisible \code{0}.
}
\description{
Set debug mode
}
\details{
When the debug mode is activated, the function \code{codestral()} will
print the request body and the response body.
}
