% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_package.R
\name{detect_package}
\alias{detect_package}
\title{Detect whether the working directory is that of a package}
\usage{
detect_package()
}
\value{
TRUE if the working directory is that of a package, FALSE otherwise
}
\description{
Detect whether the working directory is that of a package
}
