% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENDPOINTS.R
\docType{data}
\name{ENDPOINTS}
\alias{ENDPOINTS}
\title{Endpoints for the Codestral API.}
\format{
A named list with elements chat and completion.
}
\usage{
ENDPOINTS
}
\description{
Endpoints for the Codestral API.
}
\keyword{datasets}
