% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_addin.R
\name{insert_addin}
\alias{insert_addin}
\title{Insert the model's answer}
\usage{
insert_addin()
}
\value{
\code{0} (invisible).
}
\description{
This function inserts a Codestral FIM into the current script.
}
