% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory_Rfiles.R
\name{inventory_Rfiles}
\alias{inventory_Rfiles}
\title{This function returns a data frame with the file names and paths of all R files in the current directory.}
\usage{
inventory_Rfiles()
}
\value{
A data frame with the file names and paths of all R files in the current directory.
}
\description{
This function returns a data frame with the file names and paths of all R files in the current directory.
}
\examples{
inventory_Rfiles()
}
