% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestral-package.R
\docType{package}
\name{codestral-package}
\alias{codestral-package}
\title{codestral: Fill in you code with Codestral}
\description{
This package creates an addin that can be used to fill in the middle some R code.
}
\author{
\strong{Maintainer}: Marc Grossouvre \email{marc@grossouvre.fr}

}
\keyword{internal}
