% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestral-package.R
\docType{package}
\name{codestral-package}
\alias{codestral-package}
\title{codestral: Chat and FIM with 'Codestral'}
\description{
Create an addin in 'Rstudio' to do fill-in-the-middle (FIM) and chat with latest Mistral AI models for coding, 'Codestral' and 'Codestral Mamba'. For more details about 'Mistral AI API': \url{https://docs.mistral.ai/getting-started/quickstart/} and \url{https://docs.mistral.ai/api/}. For more details about 'Codestral' model: \url{https://mistral.ai/news/codestral}; about 'Codestral Mamba': \url{https://mistral.ai/news/codestral-mamba}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://urbs-dev.github.io/codestral/}
  \item Report bugs at \url{https://github.com/urbs-dev/codestral/issues}
}

}
\author{
\strong{Maintainer}: Marc Grossouvre \email{marcgrossouvre@urbs.fr}

Other contributors:
\itemize{
  \item  URBS company \email{contact@rubs.fr} [copyright holder, funder]
}

}
\keyword{internal}
