% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestral-package.R
\docType{package}
\name{codestral-package}
\alias{codestral-package}
\title{codestral: Chat and FIM with Codestral}
\description{
Create an addin in Rstudio to do fill-in-the-middle (FIM) and chat with latest Mistral AI models for coding, codestral and mamba.
}
\author{
\strong{Maintainer}: Marc Grossouvre \email{marc@grossouvre.fr} (URBS company [cph, fnd])

}
\keyword{internal}
