% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestral.R
\name{codestral}
\alias{codestral}
\title{Fill in the middle with Codestral}
\usage{
codestral(
  prompt,
  api_key = Sys.getenv(x = "R_CODESTRAL_APIKEY"),
  model = Sys.getenv(x = "R_CODESTRAL_MODEL"),
  temperature = as.integer(Sys.getenv(x = "R_CODESTRAL_TEMPERATURE")),
  max_tokens = Sys.getenv(x = "R_CODESTRAL_MAX_TOKENS"),
  suffix = ""
)
}
\arguments{
\item{prompt}{The prompt to complete.}

\item{api_key}{Your Codestral API key.}

\item{model}{The model to use.}

\item{temperature}{The temperature to use.}

\item{max_tokens}{The maximum number of tokens to generate.}

\item{suffix}{The suffix to use.}

\item{ENDPOINTS}{The endpoints to use.}
}
\value{
The completed text.
}
\description{
Fill in the middle with Codestral
}
\examples{
# codestral(prompt = "Once upon a time")
}
