% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestral_init.R
\name{codestral_init}
\alias{codestral_init}
\title{Initialize codestral}
\usage{
codestral_init(
  apikey = NULL,
  model = "codestral-latest",
  temperature = 0,
  max_tokens = 100
)
}
\arguments{
\item{apikey}{Your codestral API key}

\item{model}{The model to use}

\item{temperature}{The temperature to use}

\item{max_tokens}{The maximum number of tokens to generate}
}
\description{
Initialize codestral
}
